List of changes: * Relax restriction is_same to is_constructible. * Change container parameter types to const T&. *  Add overloads for 'addTimestamps'.

+++ OpenSim/Common/Test/DataTable.cpp
@@ -2585,7 +2585,7 @@ int main() {
                                          "col-four"};
     tsdt.setColumnLabels(tsdt_labels);
 
-    tsdt.addTimestamps(std::vector<long>{1, 2, 3});
+    tsdt.addTimestamps(std::vector<double>{1, 2, 3});
 
     return 0;
 }
